requirements.txt
streamlit==1.32.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
openai==0.28.1
numpy==1.26.4
opencv-python==4.9.0.80 